+++ protocols/relay/src/lib.rs
@@ -25,7 +25,6 @@
 
 mod behaviour;
 mod copy_future;
-mod multiaddr_ext;
 mod priv_client;
 mod protocol;